backend/requirements.txt
fastapi
uvicorn
pydantic[email]
google-genai
python-dotenv
motor
python-jose[cryptography]
bcrypt
python-multipart
openai
